+++ test/runTest.ts
@@ -8,13 +8,16 @@ async function main() {
 		const extensionTestsPath = path.resolve(__dirname, './suite');
 		const testWorkspace = path.resolve(__dirname, '../../test-fixtures/test-workspace');
 
-		await runTests({
-			version: '1.49.1',
-			extensionDevelopmentPath,
-			extensionTestsPath,
-			launchArgs: [testWorkspace]
-		});
+		const vsCodeVersions = ['1.49.1', '1.62'];
 
+		vsCodeVersions.forEach(async (version) => {
+			await runTests({
+				version: version,
+				extensionDevelopmentPath,
+				extensionTestsPath,
+				launchArgs: [testWorkspace]
+			});
+		});
 	} catch (err) {
 		console.error('Failed to run tests');
 		process.exit(1);
